+++ config.py
@@ -1,0 +1,3 @@
+import os
+__DATA_PATH__ = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'data', 'extrovert_introvert.csv')
+__EXPERIMENTS_PATH__ = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'methods', 'output')